+++ test/adlib.tape.js
@@ -1,6 +1,6 @@
 
 import test from 'tape';
-import adlib from '../lib/adlib';
+import adlib, { listDependencies } from '../lib/adlib';
 
 test('AdLib:: Throws if optional transform passed in', (t)=>{
   let template = {
@@ -76,6 +76,24 @@ test('Adlib::Strings:: allow info-window template to pass through', (t) => {
   t.plan(2);
   t.equal(result.value, 'red');
   t.equal(result.description, template.description);
+  t.end();
+})
+
+test('Adlib::listDependencies:: parse dependecies from a string', (t) => {
+  const template = 'Injuries: {{CRASHID}}<br />Fatalities: {{ISREPORTONSCENE}} Foo: {{CRASHID}}'
+  const deps = listDependencies(template);
+  t.equal(deps[0], 'CRASHID');
+  t.equal(deps[1], 'ISREPORTONSCENE');
+  t.equal(deps.length, 2)
+  t.end();
+})
+
+test('Adlib::listDependencies:: parse dependecies from an object', (t) => {
+  const template = {key: 'Injuries: {{CRASHID}}<br />Fatalities: {{ISREPORTONSCENE}} Foo: {{CRASHID}}'}
+  const deps = listDependencies(template);
+  t.equal(deps[0], 'CRASHID');
+  t.equal(deps[1], 'ISREPORTONSCENE');
+  t.equal(deps.length, 2)
   t.end();
 })